Preliminary checkin of in-progress additions to the gui package.
The code being checked in is an upstream push of code that I'm working on almost daily at work. Sadly,
there's numerous redundant firewalls between me and the outside world at that location, and I can only
push upstream from an off-site connection.

Included in this initial check-in:
 * JTerminal - a near drop-in replacement for JTerminalScreen, with color map, keyboard map, real-time
resizing with optional font scaling, and an Event dispatching mechanism opening up all sorts of fun ways to
extend it's behavior.
 * ColorMap interface and a very Kludgey DefaultColorMap - this class will change drastically in the
future.
 * KeyMap interface and DefaultKeymap.
 * TerminalEvent and TerminalEventListener for handling events from JTerminal.
 * DefaultTerminalModel, introducing a real honest-to-goodness MVC approach to a Terminal GUI object!